+++ scintilla/win32/ScintillaWin.cxx
@@ -15,6 +15,7 @@
 
 #include <string>
 #include <vector>
+#include <map>
 
 /* notepad2-mod custom code start */
 #if !defined(_WIN32_WINNT)
@@ -1348,8 +1349,10 @@ public:
 			if (lenMixed > utf16Mixed.size()) {
 				utf16Mixed.resize(lenMixed + 8);
 			}
-			size_t nUtf16Mixed = ::MultiByteToWideChar(65001, 0, mixed, lenMixed,
-				&utf16Mixed[0], utf16Mixed.size());
+			size_t nUtf16Mixed = ::MultiByteToWideChar(65001, 0, mixed,
+				static_cast<int>(lenMixed),
+				&utf16Mixed[0],
+				static_cast<int>(utf16Mixed.size()));
 
 			if (nUtf16Mixed == 0) {
 				// Failed to convert -> bad UTF-8
@@ -1362,11 +1365,14 @@ public:
 			}
 			int lenFlat = ::LCMapStringW(LOCALE_SYSTEM_DEFAULT,
 				LCMAP_LINGUISTIC_CASING | LCMAP_LOWERCASE,
-				&utf16Mixed[0], nUtf16Mixed, &utf16Folded[0], utf16Folded.size());
+				&utf16Mixed[0],
+				static_cast<int>(nUtf16Mixed),
+				&utf16Folded[0],
+				static_cast<int>(utf16Folded.size()));
 
 			size_t lenOut = UTF8Length(&utf16Folded[0], lenFlat);
 			if (lenOut < sizeFolded) {
-				UTF8FromUTF16(&utf16Folded[0], lenFlat, folded, lenOut);
+				UTF8FromUTF16(&utf16Folded[0], lenFlat, folded, static_cast<int>(lenOut));
 				return lenOut;
 			} else {
 				return 0;
@@ -1393,8 +1399,10 @@ public:
 			if (lenMixed > utf16Mixed.size()) {
 				utf16Mixed.resize(lenMixed + 8);
 			}
-			size_t nUtf16Mixed = ::MultiByteToWideChar(cp, 0, mixed, lenMixed,
-				&utf16Mixed[0], utf16Mixed.size());
+			size_t nUtf16Mixed = ::MultiByteToWideChar(cp, 0, mixed,
+				static_cast<int>(lenMixed),
+				&utf16Mixed[0],
+				static_cast<int>(utf16Mixed.size()));
 
 			if (nUtf16Mixed == 0) {
 				// Failed to convert -> bad input
@@ -1407,7 +1415,10 @@ public:
 			}
 			int lenFlat = ::LCMapStringW(LOCALE_SYSTEM_DEFAULT,
 				LCMAP_LINGUISTIC_CASING | LCMAP_LOWERCASE,
-				&utf16Mixed[0], nUtf16Mixed, &utf16Folded[0], utf16Folded.size());
+				&utf16Mixed[0],
+				static_cast<int>(nUtf16Mixed),
+				&utf16Folded[0],
+				static_cast<int>(utf16Folded.size()));
 
 			size_t lenOut = ::WideCharToMultiByte(cp, 0,
 				&utf16Folded[0], lenFlat,
@@ -1416,7 +1427,7 @@ public:
 			if (lenOut < sizeFolded) {
 				::WideCharToMultiByte(cp, 0,
 					&utf16Folded[0], lenFlat,
-					folded, lenOut, NULL, 0);
+					folded, static_cast<int>(lenOut), NULL, 0);
 				return lenOut;
 			} else {
 				return 0;
@@ -1471,7 +1482,8 @@ std::string ScintillaWin::CaseMapString(const std::string &s, int caseMapping) {
 
 	UINT cpDoc = CodePageOfDocument();
 
-	unsigned int lengthUTF16 = ::MultiByteToWideChar(cpDoc, 0, s.c_str(), s.size(), NULL, 0);
+	unsigned int lengthUTF16 = ::MultiByteToWideChar(cpDoc, 0, s.c_str(),
+		static_cast<int>(s.size()), NULL, 0);
 	if (lengthUTF16 == 0)	// Failed to convert
 		return s;
 
@@ -1486,7 +1498,7 @@ std::string ScintillaWin::CaseMapString(const std::string &s, int caseMapping) {
 
 		// Change text to UTF-16
 		std::vector<wchar_t> vwcText(lengthUTF16);
-		::MultiByteToWideChar(cpDoc, 0, s.c_str(), s.size(), &vwcText[0], lengthUTF16);
+		::MultiByteToWideChar(cpDoc, 0, s.c_str(), static_cast<int>(s.size()), &vwcText[0], lengthUTF16);
 
 		// Change case
 		int charsConverted = ::LCMapStringW(LOCALE_SYSTEM_DEFAULT, mapFlags,
@@ -1497,12 +1509,12 @@ std::string ScintillaWin::CaseMapString(const std::string &s, int caseMapping) {
 
 		// Change back to document encoding
 		unsigned int lengthConverted = ::WideCharToMultiByte(cpDoc, 0,
-			&vwcConverted[0], vwcConverted.size(),
+			&vwcConverted[0], static_cast<int>(vwcConverted.size()),
 			NULL, 0, NULL, 0);
 		std::vector<char> vcConverted(lengthConverted);
 		::WideCharToMultiByte(cpDoc, 0,
-			&vwcConverted[0], vwcConverted.size(),
-			&vcConverted[0], vcConverted.size(), NULL, 0);
+			&vwcConverted[0], static_cast<int>(vwcConverted.size()),
+			&vcConverted[0], static_cast<int>(vcConverted.size()), NULL, 0);
 
 		return std::string(&vcConverted[0], vcConverted.size());
 
@@ -1512,7 +1524,8 @@ std::string ScintillaWin::CaseMapString(const std::string &s, int caseMapping) {
 
 		// Change text to UTF-16
 		wchar_t vwcText[shortSize];
-		::MultiByteToWideChar(cpDoc, 0, s.c_str(), s.size(), vwcText, lengthUTF16);
+		::MultiByteToWideChar(cpDoc, 0, s.c_str(), static_cast<int>(s.size()),
+			vwcText, lengthUTF16);
 
 		// Change case
 		int charsConverted = ::LCMapStringW(LOCALE_SYSTEM_DEFAULT, mapFlags,
@@ -1616,8 +1629,8 @@ void ScintillaWin::InsertPasteText(const char *text, int len, SelectionPosition 
 			// add the newline if necessary
 			if ((len > 0) && (text[len-1] != '\n' && text[len-1] != '\r')) {
 				const char *endline = StringFromEOLMode(pdoc->eolMode);
-				pdoc->InsertString(insertPos + len, endline, strlen(endline));
-				len += strlen(endline);
+				pdoc->InsertString(insertPos + len, endline, static_cast<int>(strlen(endline)));
+				len += static_cast<int>(strlen(endline));
 			}
 			if (sel.MainCaret() == insertPos) {
 				SetEmptySelection(sel.MainCaret() + len);
@@ -2475,7 +2488,7 @@ STDMETHODIMP ScintillaWin::Drop(LPDATAOBJECT pIDataSource, DWORD grfKeyState,
 
 		if (data && convertPastes) {
 			// Convert line endings of the drop into our local line-endings mode
-			int len = strlen(data);
+			int len = static_cast<int>(strlen(data));
 			char *convertedText = Document::TransformLineEnds(&len, data, len, pdoc->eolMode);
 			if (dataAllocated)
 				delete []data;
